feat: switch to composable executor for Ethereum
