+++ crates/ethereum/evm/Cargo.toml
@@ -18,8 +18,6 @@ reth-evm.workspace = true
 reth-primitives = { workspace = true, features = ["reth-codec"] }
 reth-revm.workspace = true
 reth-ethereum-consensus.workspace = true
-reth-prune-types.workspace = true
-reth-execution-types.workspace = true
 reth-consensus.workspace = true
 
 # Ethereum
@@ -36,6 +34,7 @@ reth-testing-utils.workspace = true
 reth-evm = { workspace = true, features = ["test-utils"] }
 reth-revm = { workspace = true, features = ["test-utils"] }
 reth-primitives = { workspace = true, features = ["secp256k1"] }
+reth-execution-types.workspace = true
 secp256k1.workspace = true
 serde_json.workspace = true
 alloy-genesis.workspace = true
