+++ crates/ethereum/node/src/evm.rs
@@ -1,6 +1,8 @@
 //! Ethereum EVM support
 
 #[doc(inline)]
-pub use reth_evm_ethereum::execute::EthExecutorProvider;
+pub use reth_evm::execute::BasicBlockExecutorProvider;
+#[doc(inline)]
+pub use reth_evm_ethereum::execute::{EthExecutionStrategyFactory, EthExecutorProvider};
 #[doc(inline)]
 pub use reth_evm_ethereum::EthEvmConfig;
